helpers: improve RunCommand, simplify GetLastError
RunCommand now returns the result as a  map[string]interface{}. The function
also checks for errors in the result and returns them as an os.Error. If an
application want to specify the result type, then the application should use
FindOne directly.

Removed the err argument from GetLastError. Although the err argument
is convenient, it's use is not obvious.